+++ src/Oro/Bundle/UIBundle/Resources/public/js/polyfills.js
@@ -1,10 +1,4 @@
-import loadModules from 'oroui/js/app/services/load-modules';
-
 const polyfills = [];
-
-if (!window.fetch) {
-    polyfills.push(loadModules('whatwg-fetch'));
-}
 
 if (!window.sleep) {
     window.sleep = async function(duration) {