+++ src/Oro/Component/TestUtils/ORM/Mocks/StatementMock.php
@@ -25,12 +25,14 @@ class StatementMock implements \IteratorAggregate, \Doctrine\DBAL\Driver\Stateme
     {
     }
 
-    public function execute($params = null)
+    public function execute($params = null): bool
     {
+        return true;
     }
 
-    public function rowCount()
+    public function rowCount(): int
     {
+        return 1;
     }
 
     public function closeCursor()